+++ CHANGELOG.md
@@ -5,12 +5,17 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- Added
+    - People bar: alphabet
+    - Duplicates
+    - Clean up
+    
 - Improvements
- - Search
+    - Search
 
 - Bug fixes
- - Search
- - File existence indicator
+    - Search
+    - File existence indicator
 
 ## [0.0.1] - 2018-06-24
 ### Added